service_dhcp: remove support for populating dns namserver from gateway resolv.conf ( OFW-4256 )
Fix DNS IP addresses configured by the customer are not used
by the client on the LAN. Still use default 8.8.8.8

BUG DESCRIPTION:
This is a regression due to the change done for bringing in
core-net-lib changes for service-dhcp

FIX DESCRIPTION:
Added a check flag for OFW to ensure that the part
of code causing the issue is not executed.

TESTING PERFORMED:
Ensured that LAN client is using the configured DNS address.

Signed-off-by: Keerthana V J(kvijayakumarjalajaku.contractor@libertyglobal.com)