RDKBDEV-1780: fix rpcserver crash
Reason for change:
rpcserver can crash when the dhcp leases file is larger than 4k and
Plume pods can stay offline when a big number of clients is connected

rpcserver can crash when the dhcp leases file is larger than 4k and
rpcclient2 "cat /nvram/dnsmasq.leases" is called. This issue can
happen if the lease file is exactly 4k in size, or if a new line
falls exactly on the 4k boundary.

Proposed fix for allow space for the \0 at the end of the buffer filled by rpcserver.

Test Procedure:
See ticket for details

Risks: None.

Signed-off-by: Hamza Ben Zekri <hbenzekri@libertyglobal.com>
Change-Id: Iea72c4098347678790b7c2050d84c50425ecc1b9