fix Device.IP.Interface.4.Enable to false, LAN still can get IP address
[BUG DESCRIPTION]
When Device.IP.Interface.4.Enable is set to FALSE, brlan0 ip address are not deleted
since a condition is not met and LAN clients are not disconnected.

[FIX DESCRIPTION]
Added a condition to teardown brlan0 instance which will delete ipv4 address and route.
Called ipv4_resync() function for ipv4-resync sysevent which will set
Device.WiFi.Radio.1.X_CISCO_COM_ApplySetting to true and disconnect wifi clients.
In Mv3, dibbler-server is not getting started for dhcpv6s_restart event because
active_lan_interfaces is NULL and IPv6 of brlan0 is in tentative state for 2 second.
so, added a sleep for 5 second.

[TESTING PERFORMED]
1) Connect a LAN client.
2) Ensure that LAN client got ip address and is able to access Internet (Ping).
3) Now set Device.IP.Interface.4.Enable to false. Soon, LAN client will get
   disconnected and can not reach internet. Even when LAN clients are reconnected,
   it will not get ip address.
4) Set Device.IP.Interface.4.Enable to TRUE and reconnect LAN client. It will
   get ip address and can reach internet.

note:
After making Device.IP.Interface.4.Enable to FALSE, brlan0 will not lose IPv6 address
because a fix from rdkb is made to ensure gateway should advertise the old prefix with
a preferred lifetime of 0 using del_addr6_flg flag.

rmcs: basic validation of the change passed, During release renew LAN
client was not getting Ipv4 address with when Ipv4 disabled on brlan0
interface.

Upstream Candidate: No.

Signed-off-by: akashdhanabal <adhanabal.ext@libertyglobal.com>