fix guest wifi DNS configuration in DS-Lite mode ( OFW-4683 )
[Issue]
When the WAN DNS configuration is not available (ie: in DS-Lite mode),
dnsmasq is configured to provide the brlan0 IP address to the LAN
clients instead. This is a problem for guest wifi clients as they
are blocked from accessing brlan0 IP.

[Solution]
Configure dnsmasq to supply their own brlanX IP address to clients
of each network separately.

[Testing]
1) Configure DS-Lite mode and enable guest wifi
2) Connect a client to private network and confirm that they receive
   brlan0 IP as their DNS configuration.
3) Connect a client to guest network and confirm that they receive
   brlan7 IP as their DNS configuration.
4) In WebUI, enable customer defined DNS servers and configure preferred
   DNS IP, leaving alternate fields empty.
5) Check both private and guest wifi networks again and confirm:
    - Clients have the configured preferred DNS as their first DNS config
    - Clients have the correct brlanX IP as their second DNS config

[Upstream Candidate]
No. LG specific

Signed-off-by: Mustafa Karadzha(mkaradzha@libertyglobal.com)