apply_system_defaults: allow setting a default syscfg value to an empty string
It's always been possible to set a syscfg value to an empty string
when calling syscfg_set() directly, however previously it was not
possible to define a default value as an empty string. There's no
clear reason why it should not be possible to do that.

This is a change from previous behaviour.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>