apply_system_defaults: don't update syscfg values previously set to an empty string
A syscfg variable should not be set to its default value if already
set to an empty string.

This is a change in behaviour: previously a syscfg variable would be
set to its default value if undefined or set to an empty string. It
will now only be set to its default value if undefined.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>