apply_system_defaults: clear ntp_server1, ntp_server2, etc on every reboot
NTP server values will be set from DHCP on every reboot (ie values
are stored in syscfg for convenience, not because they need to be
persistent across reboots). Clear values from apply_system_defaults
to avoid any risk of stale values (e.g. if the number of NTP servers
provided by DHCP is fewer that it was on a previous reboot).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>