service_dhcp: drop support for Xfinityhome interface ( OFW-4965 F5685SI-1557 )
[Issue]

On bootup the following error message is visible on the logs: "brlan1 interface is not present, cannot get IPv4 Address".

[Description]

Brlan1 interface is not even used, misleading error message is coming at bootup.

[Solution]

Remove the function call is_iface_present from dhcp_server_start which checks for the presence of brlan1.

[Testing]

Upgrade the image and checked the bootup logs.

Signed-off-by: Saravanakumar Karthikeyan <skarthikeyan.contractor@libertyglobal.com>