firewall: add source ip check before nat ( OFW-5151 )
temp_fix: no internet access on clients when single-static RIPv2 is enabled on Gateway

[DESCRIPION]
NAT port translation is now applying for all the traffics originated from the Gateway console as well,
This is causing the RIPv2 Routes to not to get added in CMTS.

According to RFC2543, All communications intended for another routers's
RIP process are sent to the RIP port.  All routing update messages are sent from the RIP port

[FIX DESCRIPTION]
skip nat for the traffics originated from erouter0

[TESTING PERFORMED]
1. Verified internet connection on LAN client when Single Static RIP is
enabled.
2. Verified internet connectivity on Gateway console.

Note:- The cleanup will be done to have a logic to mark the in host packets at
mangle table and filter accordingly

Signed-off-by: BijeeshaVS <bvattakaithayilsudha.contractor@libertyglobal.com>