service_ntpd.sh: fix DUT time status showing as unsynchronized for fiber devices
[Issue]
NTP quick sync is not working when no static NTP server is configured
in the ntp_server1 syscfg variable, even when an NTP server received
through DHCPv4/6 is available. This is a regression introduced by the
rebase to R1.7.

[Solution]
Avoid setting NTP_SERVER_MODE when the ntp_server1 syscfg variable
does not contain a URL.

[Testing Performed]
Reboot the gateway
Verify in /rdklogs/logs/ntpLog.log that an NTP quick sync was launched and succeeded
Verify that the gateway's time was updated

[Upstream Candidate]
Yes