fix GPV of Service Provider DNS servers returns empty when DNS override is enabled via ACS and Customer defined DNS servers are not configured ( OFW-5339 )
[BUG DESCRIPTION]
When the DNS Override is enabled and Customer side DNS(preferred) is not configured, the dibbler server is getting crashed.

[FIX DESCRIPTION]
Added a condition to remove trailing spaces from the beginning and ending of each ip's.
And also adding (,)comma in between the ip's, only if there are 2 or more ip's.
[Testing]
1) Enable the DNS override.
2) Service Provider side DNS server ip need to persists. When the customer side DNS servers
   are not configured and dibbler-server should be up.

Signed-off-by: vchandrakanth.ext <vchandrakanth.ext@libertyglobal.com>