avoid syslogd writing systemlog etc to /nvram ( OFW-5678 )
[Description]
An issue related to WIFI SSID defaulted to UPCXXXXX from VMXXXXX
after the FW update was being debugged on MV2+. The RC was concluded
to be the syslogd writing often into /nvram/log/system/systemlog
when a UPnP capable client is connected to the CPE. After a period
of time, the /nvram either become read-only or fails to mount when
the device is rebooted post an upgrade and the /nvram is formatted.

[Solution]
The final solution around making the non-vol resilient to such issues
is sitting with Sagemcom.

However, the immediate change to help mitigate the solution is to
move the syslogd writes into /var/log instead of writing into
nvram. A full assessment of whether the logs are required to be
persistent is yet to be done.

The changes done for F3896SI-161 also has an impact on how often the
logs are pushed into the systemlog. That also needs to be reviewed.

[Testing Performed]
1. Boot up the gateway.
2. Once the device is online, enable DMZ or perform UPnP tests
3. Ensure that the logs are written into /var/log and not into nvram

To be backported to R1.7 : Yes

Signed-off-by: Ajay Vishnu Bhat <abhat.ext@libertyglobal.com>