RDKBDEV-2510 : Unchecked return value.
Reason for Change : If the function returns an error value, the error value may be mistaken for a normal value.

In daemon_init function of trigger.c and syseventd_main.c files: Value returned from the chdir() function is not checked for errors before being used.

Test Procedure :
1.Not Applicable

Risks : None

Change-Id: I83e5b7dc38af6fcf18d51dcf8b2c97d0110632f2
Signed-off-by: ajithsagar.z@infosys.com