RDKBDEV-2581 : Branch Past Initialization
Reason for Change :  Issue related to the scope of variables val and out_match. These variables are declared inside a switch statement but after a label which lead to “jumping over initializations” issue.

The proposed solution is to move the declarations of val and out_match before the switch statement which maintains security and proper code execution.

Test Procedure :
1.Not Applicable

Risks : None

Change-Id: I7cc435ee90e61d908fbe06c46693467b85c89ce3
Signed-off-by: ajithsagar.z@infosys.com