RDKBDEV-2584 : Parse Warning
Reason for Change :The issue is related to the variable id. The function parameter id is being hidden by local variables of the same name inside the SE_INC_LOG macro calls. This can lead to confusion and bugs because changes to the id inside the macro will not affect the id outside, and vice versa.

Test Procedure :
1.Not Applicable

Risks : None

Change-Id: If63203c1e6ef638b3745013f00c36e626babe5cb
Signed-off-by: ajithsagar.z@infosys.com