fix block fragmented IP packets firewall functionality is working when VMB is enabled ( OFW-5808 F3896SI-2647 )
[Issue Description]
Block fragmented IP packets firewall functionality is working and
ping to eRouter WAN IP with packet size greater than MTU size from
LAN fails when VMB is enabled. As per MVXREQ-2407, support for
firewall functionality should be turned off. This is because while
disabling firewall functionality during Multi-static VMB/RIPV2 is
enabled, the call to handle the "Block fragmenetd IP packets"
functionality is not being handled.

[Fix Description]
Added call to do_blockfragippktsv4() and in that method added a check
if any of the multi -static solution is enabled. If so unblock the
fragmented IP packets.

[Testing performed]
1. Enable Device.Firewall.X_RDKCENTRAL-COM_Security.V4.BlockFragIPPkts to true.
2. Make sure the packets with greater MTU size is blocked. (i.e ping <erouter_ip> -c 3 -s 1800)Ping fails.
3. ENable MS VMB.
4. The packets with greater MTU size is not blocked.
5. Repeat same steps 1 to 4 for RIPV2 case as well

Signed-off-by: Keerthana V J(kvijayakumarjalajaku.contractor@libertyglobal.com)

firewall: misc code cleanup ( Part of OFW-5808 )

Signed-off-by: BijeeshaVS <bvattakaithayilsudha.contractor@libertyglobal.com>