firewall: fix packet loss observed in ICMP ping test ( OFW-5856 )
[Issue]
ICMP echo requests are not receiving a reply when sent from a
LAN/WLAN client. It's related to NAT randomization being applied to
the ICMP Echo Request ID. For the first ICMP requests, the ID is sent
on erouter0 with the value 0. In this case, the server replies with a
completely different ID and the gateway is not able to NAT it back to
the original client. If two ping sessions are started at the same
time, then the ID will increment and future pings will succeed.

[Solution]
Remove randomization when NATting ICMP messages (fix taken from OFW-5294).

[Testing performed]
Reboot the gateway
Run "ping 8.8.8.8" from a LAN client
Verify that it succeeds