fix NAT hairpin functionality not working ( OFW-5844 )
[BUG DESCRIPTION]
A host on the LAN side of the modem is unable to reach a web-server on the same LAN side by
using the erouter0 ip with the correct port-forwarding configured.

[FIX DESCRIPTION]
After setting up nat hairpin, iptables DNAT rules will be executed prerouting chain of bridge forwarding (before bridging code executed)
and the destination MAC address will be modified causing preventing the reception on brlan0 due to MAC address mismatch.

In order to Prevent this brlan0 interface is moved to promiscuous mode, so that  it passes all network traffic it receives,
rather than only the frames specifically addressed to it.

Legacy mv1 device also has promiscuous mode for the corresponding lan side interface.

[TESTING PERFORMED]
1. Bring the gateway online.
2. Confirmed that NAT hairpin functionality is working fine.