avoid GenFWLog from writing the specific firewall logs into /nvram ( OFW-5685 )
[Issue]
Avoid GenFWLog from writing the specific firewall logs into /nvram

[Description]
As part of efforts to limit the writes to nvram, the logs generated by
GenFWLog can be moved to /var/logs instead of /nvram/log/firewall

[Solution]
The immediate change to reduce the number of writes into nvram is by
redirecting the GenFWLog to write into /var/log instead. A full
assessment of whether the logs are required at all is still to be assessed.

[Testing]
1. Boot up the gateway.
2. Once the device is online, check if the fwlog.tar.bz2 is created in
   /var/logs

[Upstream Candidate]
Maybe

Signed-off-by: Jyoti <jjyoti@libertyglobal.com>