utopia_init.sh: more robust parsing of Broadcom specific /proc/device-tree/bolt/reset-list ( OFW-5819 F3896SI-2654 )
Fix power off reboot reason getting reported as unknown ( OFW-5819 F3896SI-2654 )

[Issue Description]
To upstream the patch - sagemcom-fix-main-chip-input-reboot-type-reason-is-unknown.patch done to fix the issue F3896SI-2593

[Solution]
Upstreamed the patch as generic code across the platform.

[Testing]
Ensured that the changes works fine for MV2+ and it doesn't impact the flow of other platforms.

Signed-off-by: Keerthana V J(kvijayakumarjalajaku.contractor@libertyglobal.com)