utopia_init.sh: reboot event due to SW upgrade is not displayed in Log file ( OFW-5556 F3896SI-1544 )
[Issue Description]
ccsp-p-and-m updates the value of param X_RDKCENTRAL-COM_LastRebootReason
by checking the presence of file /var/tmp/lastrebootreason.

If this file does not exist, and the reboot counter is 0,
there will be such trace in file /rdklogs/logs/BootTime.log : Received reboot reason as unknown.
For this reason, this param has always been unknown in former releases.

[Solution]
Once there is a reboot caused by SU, utopia shall change X_RDKCENTRAL-COM_LastRebootCounter to 1,
and change X_RDKCENTRAL-COM_LastRebootReason value to "Software_Upgrade".
Thus, ccsp-p-and-m shall not set it to "unknown"

[Testing Performed]
1.  Perform a software upgrade, wait until the gateway fully boots up;
    then perform again a software upgrade

2.  Check syscfg param of RebootReason
    syscfg show | grep -i RebootReason
    output: X_RDKCENTRAL-COM_LastRebootReason=Software_Upgrade

3.  Check the value of DM Device.DeviceInfo.X_RDKCENTRAL-COM_LastRebootReason
    dmcli eRT getv Device.DeviceInfo.X_RDKCENTRAL-COM_LastRebootReason
    Its value should be "Software_Upgrade"

4.  Wait for about 20 minutes, check if this reboot reason have been logged in SelfHeal log
    cat /rdklogs/logs/SelfHeal.txt.0 | grep -i reboot
    log line LastrebootReason = Software_Upgrade should be present

Signed-off-by: venkat (vturaka.ext@libertyglobal.com)