temp_fix: NTP (IPv4) seems to be blocked by gateway after upgrade to R2.0RC3 (R2.1) ( OFW-6015 F3896SI-2717 )
[Issue Description]
With the NAPT changes there are regressions found in R1.7
( F3896SI-2434 ) as well as in R2.0 ( F3896SI-2717 ). Hence decided
to revert the NAPT changes and requirement team will review the
MVXREQ-1091 to understand if this is really need to be addressed with
the known regressions it can cause.

[Solution]
Reverted NAPT implementation. (Applicable to all platforms)

[Testing Performed]
NAPT is not observed while trying to access IPV4 site from LAN client (excpet for address)

Signed-off-by: Keerthana V J(kvijayakumarjalajaku.contractor@libertyglobal.com)